[risk=low][RW-8309] Move WorkspaceResourceService and UserRecentResourceService to workspaces/resources (#6850)
