+++ api/src/main/java/org/pmiops/workbench/api/ConceptSetsController.java
@@ -8,7 +8,6 @@ import java.util.stream.Collectors;
 import javax.inject.Provider;
 import org.apache.commons.collections4.CollectionUtils;
 import org.pmiops.workbench.conceptset.ConceptSetService;
-import org.pmiops.workbench.db.dao.UserRecentResourceService;
 import org.pmiops.workbench.db.model.DbUser;
 import org.pmiops.workbench.db.model.DbWorkspace;
 import org.pmiops.workbench.exceptions.BadRequestException;
@@ -20,6 +19,7 @@ import org.pmiops.workbench.model.EmptyResponse;
 import org.pmiops.workbench.model.UpdateConceptSetRequest;
 import org.pmiops.workbench.model.WorkspaceAccessLevel;
 import org.pmiops.workbench.workspaces.WorkspaceAuthService;
+import org.pmiops.workbench.workspaces.resources.UserRecentResourceService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RestController;
