+++ api/src/main/java/org/pmiops/workbench/workspaces/resources/WorkspaceResourcesService.java
@@ -1,4 +1,4 @@
-package org.pmiops.workbench.cdrselector;
+package org.pmiops.workbench.workspaces.resources;
 
 import java.util.List;
 import org.pmiops.workbench.db.model.DbWorkspace;
@@ -9,6 +9,6 @@ import org.pmiops.workbench.model.WorkspaceResource;
 // This is a service for accessing objects that describe selections of CDR information. For
 // example, datasets and cohorts, both of which define subsets of the CDR.
 public interface WorkspaceResourcesService {
-  public List<WorkspaceResource> getWorkspaceResources(
+  List<WorkspaceResource> getWorkspaceResources(
       DbWorkspace dbWorkspace, WorkspaceAccessLevel accessLevel, List<ResourceType> resourceTypes);
 }
