+++ api/src/test/java/org/pmiops/workbench/workspaces/resources/UserRecentResourceServiceTest.java
@@ -1,4 +1,4 @@
-package org.pmiops.workbench.db.service;
+package org.pmiops.workbench.workspaces.resources;
 
 import static com.google.common.truth.Truth.assertThat;
 
@@ -13,8 +13,6 @@ import org.pmiops.workbench.db.dao.CohortDao;
 import org.pmiops.workbench.db.dao.ConceptSetDao;
 import org.pmiops.workbench.db.dao.UserDao;
 import org.pmiops.workbench.db.dao.UserRecentResourceDao;
-import org.pmiops.workbench.db.dao.UserRecentResourceService;
-import org.pmiops.workbench.db.dao.UserRecentResourceServiceImpl;
 import org.pmiops.workbench.db.dao.WorkspaceDao;
 import org.pmiops.workbench.db.model.DbCohort;
 import org.pmiops.workbench.db.model.DbConceptSet;
